Split `ModuleSession` to different file
